economist/requirements_ui.txt
# Streamlit Web 界面依赖
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
pyyaml>=6.0




